+ added shell script to clear the cache

+++ class/CIA.php
@@ -927,6 +927,9 @@ class
 
 	static function writeWatchTable($message, $file, $exclusive = true)
 	{
+		$file = realpath($file);
+		if (!file) return;
+
 		$file =  "++\$i;unlink('" . str_replace(array('\\',"'"), array('\\\\',"\\'"), $file) . "');\n";
 
 		foreach (array_unique((array) $message) as $message)
